requirements.txt
Django==5.0.6
djangorestframework==3.15.2
rapidfuzz==3.10.0
geopy==2.4.1
requests==2.32.3